chore: add more methods to TypeInterface that are needed by other systems
